Re-organize flow dependency modals
