+++ temba/orgs/views/views.py
@@ -145,84 +145,6 @@ class IntegrationFormaxView(IntegrationViewMixin, ComponentFormMixin, SmartFormV
         return response
 
 
-class DependencyModalMixin(OrgObjPermsMixin):
-    dependent_order = {"campaign_event": ("relative_to__name",), "trigger": ("trigger_type", "created_on")}
-    dependent_select_related = {"campaign_event": ("campaign", "relative_to")}
-
-    def get_dependents(self, obj) -> dict:
-        dependents = {}
-        for type_key, type_qs in obj.get_dependents().items():
-            # only include dependency types which we have at least one dependent of
-            if type_qs.exists():
-                type_qs = type_qs.order_by(*self.dependent_order.get(type_key, ("name",)))
-
-                type_select_related = self.dependent_select_related.get(type_key, ())
-                if type_select_related:
-                    type_qs = type_qs.select_related(*type_select_related)
-
-                dependents[type_key] = type_qs
-        return dependents
-
-
-class DependencyUsagesModal(DependencyModalMixin, SmartReadView):
-    """
-    Base view for usage modals of flow dependencies
-    """
-
-    slug_url_kwarg = "uuid"
-    template_name = "orgs/dependency_usages_modal.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["dependents"] = self.get_dependents(self.object)
-        return context
-
-
-class DependencyDeleteModal(DependencyModalMixin, ModalMixin, SmartDeleteView):
-    """
-    Base view for delete modals of flow dependencies
-    """
-
-    slug_url_kwarg = "uuid"
-    fields = ("uuid",)
-    submit_button_name = _("Delete")
-    template_name = "orgs/dependency_delete_modal.html"
-
-    # warnings for soft dependencies
-    type_warnings = {
-        "flow": _("these may not work as expected"),  # always soft
-        "campaign_event": _("these will be removed"),  # soft for fields and flows
-        "trigger": _("these will be removed"),  # soft for flows
-    }
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        # get dependents and sort by soft vs hard
-        all_dependents = self.get_dependents(self.object)
-        soft_dependents = {}
-        hard_dependents = {}
-        for type_key, type_qs in all_dependents.items():
-            if type_key in self.object.soft_dependent_types:
-                soft_dependents[type_key] = type_qs
-            else:
-                hard_dependents[type_key] = type_qs
-
-        context["soft_dependents"] = soft_dependents
-        context["hard_dependents"] = hard_dependents
-        context["type_warnings"] = self.type_warnings
-        return context
-
-    def post(self, request, *args, **kwargs):
-        obj = self.get_object()
-        obj.release(request.user)
-
-        messages.info(request, self.derive_success_message())
-        response = HttpResponse()
-        response["Temba-Success"] = self.get_success_url()
-        return response
-
-
 class LoginView(Login):
     """
     Overrides the smartmin login view to redirect users with 2FA enabled to a second verification view.
